requirements.txt
Faker==37.1.0
mysql-connector-python==9.2.0
paramiko==3.5.1
python-dotenv==1.1.0
